+++ client/src/components/ProductCard.js
@@ -5,7 +5,7 @@ import { useNavigate } from "react-router-dom";
 import "./css/ProductCard.css";
 import StarIcon from "../images/StarIcon";
 function PrintStars(props) {
-  const num = parseInt(props.number);
+  const num = parseFloat(props.number);
   if (num > 4) {
     return <div><StarIcon /><StarIcon /><StarIcon /><StarIcon /><StarIcon /></div>
   } else if (num > 3) {
